Move cacher whitebox test utils out of test file
